+++ src/parser/block/paragraph/mention/quote.test.ts
@@ -40,6 +40,7 @@ describe('Unit: parser/block/paragraph/mention/quote', () => {
       assert.deepStrictEqual(inspect(parser('> >>0\n> > b\n> c')), [['<span class="quote">&gt; <a class="anchor" href="?res=0">&gt;&gt;0</a><br>&gt; &gt; b<br>&gt; c</span>'], '']);
       assert.deepStrictEqual(inspect(parser('> > a\n> > b\n> > c')), [['<span class="quote">&gt; &gt; a<br>&gt; &gt; b<br>&gt; &gt; c</span>'], '']);
       assert.deepStrictEqual(inspect(parser('> > > a\n> > > b')), [['<span class="quote">&gt; &gt; &gt; a<br>&gt; &gt; &gt; b</span>'], '']);
+      assert.deepStrictEqual(inspect(parser('>$a-b$')), [['<span class="quote">&gt;$a-b$</span>'], '']);
       assert.deepStrictEqual(inspect(parser('>${a}$')), [['<span class="quote">&gt;<span class="math notranslate" data-src="${a}$">${a}$</span></span>'], '']);
     });